pi navving

+++ Software/STM32/thikc/Src/pid.c
@@ -1,7 +1,7 @@
 #include"pid.h"
 #include"stm32f1xx.h"
 
-PID_t pid_Init(int kp, int kd, int ki, int i_max, int g){
+PID_t pid_Init(uint8_t kp, uint8_t kd, uint8_t ki, int8_t i_max, uint8_t g){
     PID_t PID;
     PID.kp = kp;
     PID.kd = kd;
@@ -15,11 +15,11 @@ PID_t pid_Init(int kp, int kd, int ki, int i_max, int g){
     return PID;
 }
 
-double pid_GetGain(PID_t *PID){
+int16_t pid_GetGain(PID_t *PID){
     /* Get current values for P, I, and D */
     PID->count += 10;
-    double p = PID->kp * PID->err;
-    double d = PID->kd * PID->err / (PID->count);  // / 1000 to get seconds
+    int16_t p = PID->kp * PID->err;
+    int16_t d = PID->kd * PID->err / (PID->count);  // / 1000 to get seconds
     PID->i = PID->i + PID->ki * PID->err;
     
     /* Cap absolute value of integral control to keep it stable */
